Eliminate the conflict between SIGNED_INT and NUMBER
NUMBER matches all integers, but so does SIGNED_INT. We need to be able to match only signed integers for list indexes and slices, and we can't use NUMBER for that since it would allow `foo[1.2]` or `foo[2e13]`.

This removes integers from the NUMBER rule and renames it to the perhaps over-descriptive REAL_OR_EXPONENT_NUMBER. SIGNED_INT is added directly in the json rule.

Perhaps we'll need to add labels in the json rule to be able to match the different alternatives, but we'll do that when it becomes necessary.