Add support for JSON literals
The whole JSON fragment is stored as a string even though we have the parse tree. I'm not sure we want the parsed structure, since we won't be able to compare it in a generic fashion, i.e. how to compare the JSON with something else will depend on that something else. If the input is Jackson nodes the JSON will probably need to be parsed with Jackson to be comparable.