Fix bad assumptions when handling implicit current nodes
Expressions with brackets come in two forms, in the grammar I've called them "bracket expression" and "bracketed expressions", the former is when there's no expression in front of the bracket, e.g. `[1]` vs. `foo[1]`.

You could say that the former has an implicit current node in front of it where the latter has an expression.

I assumed, wrongly, that I could detect which of the case we were in by checking if the stack was empty, but that only works when the brackets are in a unary expression or on the left hand side of a binary expression – not on the right hand side, e.g `foo | [*]`.

Instead of checking for an empty stack we need to make sure we know which context we're in. If we're in a bracket expression we shouldn't look for any expression, but if we're in a bracketed expression we should.

I'm starting to think that the grammar probably should change so that these two cases will have different listener methods, but I'm unsure how at the moment.