Completely rewrite the AST generation
The current AST is inconsistent. Some constructions are left-right and some are hierarchical. When I started writing an interpreter I realized that projections are going to be a pain in this model.

This AST is only hierarchical. Each node (with some exceptions) has a source which represents the query to the left. Exceptions are nodes like AndNode and OrNode which have two branches that both have the query to the left as their source – these get the equivalent of @ as their source and the interpreter will have to sort that out. Also nodes like StringNode, JsonLiteralNode, etc. that are literal values don't have sources.