Implement support for @ (current node)
With this I realized that I had made some bad assumptions about how @ works: there are at least two different @'s, the implicit current-node operator used in for example function calls, but also the explicit current-node which works as a no-op in pipe expressions. The latter needs to refer back to the source expression and act as a pass-through. A later optimization pass could remove these nodes.