Build projection handling into JmesPathNode
Instead of having the separate ProjectionNode class.

This required fixing the case where null is passed into a projection – it should not result in an empty projection but null.

It also required changing a few test cases around negation that were wrong. `![?…]` doesn't do what you expect since `[?…]` starts a projection, `[?…] | !@` is probably the right syntax to use.