Add the basics for function support including expression references
The idea is to make functions generic in the same way as the interpreter/AST is generic. Function should be able to do their job using only the same adapter as the interpreter uses. The adapter has a reference to a function registry (of which there will be a default that will come loaded with the built-in functions), so that in the future you will be able to register your own functions.

Expression references make everything more complicated since arguments to functions can't be just List<T>. I guess the arguments could be passed as List<JmesPathNode> to the function if the current node was also passed, but that complicates things in other ways. The current solution is to have a union type ExpressionOrValue which can carry either an expression or a T.

No support for type or arity checking is currently implemented.