Add a Java Collections Framework adapter
It uses the JMESPath parser as a JSON parser.

This required adding Adapter#createNumber, and the use of LinkedHashMap instead of HashMap to preserve ordering of objects. Neither JSON nor JMESPath mandates any order of properties of an object, but Jackson preserves ordering, and it will be so much easier to test things if the order is defined.