Split the project into two modules: core and jackson
We want to add new adapters, but we don't want to have a JRuby adapter depend on Jackson, say. By having a separate module for each adapter we can avoid the dependencies snowballing.

I've removed the jp tool, because it depended on Jackson, and it wasn't really working anyway.

I've also had to copy the generic AdapterTest into both modules, I'll try to figure out how to avoid that – that's definitely not something that we want.

The compliance runner is also in the jackson module since it uses Jackson for parsing JSON. I'll see if we can move it back to core, or what we'll do about that.