Change #toDouble():Double -> #toNumber():Number and add #createNumber(long):T
Just because the primary case is to process JSON that doesn't mean we can assume everyone wants to treat all numbers as doubles. This changes the API so that the adapter will not have to convert numbers (toDouble -> toNumber) and functions like length will not have to return their values as doubles.

Functions like abs, ceil and floor still convert to doubles (the alternative for abs would be to check the input type, and it's doubtful that it's worth it); sort will convert non-doubles to doubles before comparing them.