Change the Adapter API to (hopefully) avoid conversions
* #getProperty(T,String) gets a sibling #getProperty(T,T) that can be used for no-conversion lookups in the merge function
* #getPropertyNames now returns Collection<T> instead of Collection<String>, which simplifies the keys function
* #createObject now takes a Map<T,T> instead of Map<String,T>

I'm torn as to whether this is a good change or not. With the current adapters it's worse, because Jackson needs more conversions now than before, or at least more code (for example: you don't get StringNode objects when you ask for an object's keys, you get strings and creating an object requires using strings as keys, not StringNode).

The upside is when we start creating adapters for JRuby and POJOs, and things that don't map as well to JSON. In JRuby the keys of hashes will not be String but RubyString, and they are not cheaply converted to String).

It would be nice to be able to get rid of getProperty(T,String), but I don't see that happening since PropertyNode can be created without an adapter being involved.