Move all type checking to a type checker DSL
Instead of each function checking its arguments' types in its own way this adds a type checker DSL. This greatly simplifies the function implementations.

The @Function annotation has been removed since it wasn't possible to add this DSL to it – that would have been ideal, but it's just not possible.