Add JmesPathExpression#search
This replaces #evaluate, and changes how expressions are created and used. Instead of making it possible to use the same object with multiple runtimes the expression object now keeps a reference to the runtime. The benefit of the old behaviour was very small, and the new behaviour is so much easier to use. There's no longer any need to pass in a runtime when searching a structure.

This means that JmesPathExpression is now JmesPathExpression<T>, and that you no longer create expressions through JmesPathExpression.fromString but through JmesPathRuntime#compile.