Hide the runtime internals by re-introducing the Adapter interface
It looks pretty messy with all those internal methods in the runtime interface. This removes everything that a user of a runtime doesn't need to know and puts it back into the Adapter interface, which now contains the pieces that only an implementer of a runtime needs to worry about.