Change ValueOrExpression to contain Expression and not Node
It just makes sense, it's in the name.

This also changes the implementation of ValueOrExpression. I thought it looked a bit messy determining whether or not something is a value by whether or not it has an expression. This adds two subclasses that have the two specializations. They are created with constructor methods called "of", which seems to be the craze in the Java world lately.