Make sure the keys function always returns a list
JcfRuntime#createArray should not return a Set, even though it accepts Collection as JSON array. The reason is that the JSON parser will create lists and for example [?keys(foo) == `[1, 2, 3]`] would fail if one side produced a set and the other a list.

Also fixed a bad cast in JcfRuntime#isTruthy