Reimplement SliceNode
This changes the core of the SliceNode logic to perform one simple loop, where elements are successively added to an ArrayList. The number of elements is precomputed as to avoid intermediate allocations of the array list.

The previous logic was rewritten using ?: expressions in order to highlight the similarities (and differences) between start and stop.
If we ignore the default values for start and end for a moment, it turns out that the only difference between forward and backward iteration is that the lower limit for then end for backward iteration is -1 and not
zero.

In order to initialize the default values, forward iteration is simple, as we can just use 0 and `Integer.MAX_VALUE` as the defaults. For backward iteration, we can use -1 as the start, as that will always be translated to `elements.size() - 1`, but the stop is a bit tricker as we would need a value N such that `elements.size() + N` is always negative. Thanks to two-complement representation of integers `Integer.MIN_VALUE` fits the bill, as it happens to be more negative than `Integer.MAX_VALUE` is positive, and we cannot have a `List<T>` that is longer than `Integer.MAX_VALUE`.

I decided to inline handling of default values in the constructor, which  I guess technically changes equals so that `SliceNode(runtime, 0, 10, 1, source)` is equal to `SliceNode(runtime, null, 10, 1, source)`, but they will behave in the same way, so I guess it doesn't really matter.

It also affects the toString method, as it will not print the computed default values instead of `null` for those cases. I guess both fields could be retained if this identity is important.

I also removed the protected accessor methods, as they only seemed to be used inside the class, and there it seems to make more sense to use the fields directly.