Fix SliceNode rewrite
I was sure the tests were passing, so I'm not really sure what happened.

It turns out (unsurprisingly) that also the begin range needs to be adjusted for backwards traversal (to avoid `IndexOutOfBoundsException`).

Also, the iteration logic required the `begin + N * step == end`, which is obviously not always true, and would omit the last element if this wasn't the case. While the end could certainly be adjusted to make everything fit, I thought it was nicer to use the number of items already computed for the ArrayList allocation. When making that change, I quickly realized that this computation is also wrong when step-sizes do not align, so rounding was added to ensure the final step is included.