Reimplement projections
The current implementation of projections only handles one level of projections. The projectionLevel thing in Node was an attempt to handle multiple levels, but it broke down.

This is not a great solution, but it works. What happens is that the old StartProjection and StopProjection(s) nodes now only are temporary. They serve as the delimiters for a second pass over the tree that rewrites it so that projections are correctly delimited.

This second pass is necessary because the semantics of projections are not reflected in the syntax of JMESPath. Minus some details `foo[*].bar` should be interpreted like `map(&bar, foo)` and `foo[*].bar[*].baz` as `map(&map(&baz, bar), foo)`, and this is complicated when you get to `foo[*].bar[*] | [1].baz` because the pipe stops all projections and there is no parser rule that captures this.

It doesn't strictly have to be a second pass, it could probably be weaved into the existing parser logic, but it was difficult enough to figure out how to rewrite the tree when you could see the whole tree that I didn't think I would manage that. Maybe that will come later.

One surprise that I got at the end of this was that I realized that `foo[*].bar[]` really is `(foo[*].bar)[]` or `foo[*].bar | []`. That is, the flatten operator stops projections, flatten, and then start them again. This is not really mentioned in the spec, and not at all obvious from the grammar.

It is becoming clear that the current AST is not optimal. That it's kind of upside down is making it difficult to reason about when you need to rewrite it like this. I had an AST in the other direction from the beginning, but changed because I ran into trouble with projections – but I think I misinterpreted what the problem was.