Handle missing functions better
For a short while, during parsing there exist FunctionCallNodes with no implementation. We don't bail directly when we encounter a function call to a non-existent function, just like we don't bail at the first parse error. This means that FunctionCallNode should handle that its implementation is null, because you might use hashCode, or toString during parsing.