Reorganize AST for top-down evaluation
This was mainly implemented as a proof-of-concept, and especially the implementation of ExpressionParser is a bit not-so-nice.

I tried to stay away from optimizing the structure, and while I originally intended to flatten the Sequence nodes to long chains instead of always having two children, I reverted to the somewhat simpler implementation first.

It turns out to be marginally less code.