Require errors to contain the expected words
To be able to use the jp-compliance tool in jmespath.test error messages must contain the words from the compliance suites. When a test has an expected error "syntax" the error message needs to contain the word "syntax", and when the expected error is "invalid-arity" the error message needs to contain either that string exactly or "invalid arity" (with a space instead of hyphen).