Add focused tests to increase test coverage
The test coverage is almost complete, and except for practically unreachable statements (like default cases in switches that cover all possible cases) I found a few lines {Jcf,Jackson,Base}Runtime that weren't covered. The reason is in most of these cases that there's a check in the adapter method for what should happen when it's called with an unexpected value type, but all callers check the type of the value before calling the adapter method – for example: only functions call Adapter#toNumber, but functions already check that their arguments are numbers. I still think it makes sense for #toNumber to check that it's not called with a non-number, because in the future it might be called from somewhere that does not check, and we want existing adapters to work with that.