Optimize PropertyNode and deprecate Adapter#getProperty(T,String)
Converting strings from T to String is a big performance problem, and this eliminates one of the biggest culprits. Adapter#getProperty(T,String) is no longer used after this change, so it's deprecated.

This has no impact at all for the JUC runtime, but possibly a small negative impact for the Jackson runtime, since it's JsonNode#get expects a string and not a JsonNode. Hopefully StringNode#textValue doesn't do any conversions. For the upcoming JRuby implementation this is a huge performance optimization though, since it doesn't use Java strings internally.