Compare the raw JSON values of JSON literals
To be consistent with how the other node types are compared. Node#equals is really only used for the parser tests, so it's not a big deal. Unfortunately some tests of unescaping were in the parser tests and should be in the runtime tests instead.

Since I touched all of the parser tests I took the opportunity to rewrite them as specdoc-esque sentences.