Check function arity at compile time
We have all the information we need to check whether or not a function is called with the correct number of arguments at compile time. Doing this will save us from having to deal with runtime errors.

The runtime checking of arity still remains, but the plan is to remove it.

The error message for ArityException has been changed to work better as a part of a ParseException error message, and the code that constructs ArityException messages is used by the compiler to create the ParseException message.

I took the time to rewrite all default function tests so that they don't use `@Test(expected = …)` but instead catch the exception (or fail if no exception is thrown) and assert on the error message. Perhaps it's overkill to have tests for the arity and argument types of every single default function, but we have them so we might as well keep them.

I added some additional tests to ParserTest to cover that arities are checked at parse/compile time.