Merge CompareByFunction & SortByFunction
Their implementations are very similar, too similar to be acceptable. At the same time it's not obvious how to merge them since one reduces the input to a single element and the other just reorders the input.

I made an attempt to add abstract methods directly to the CompareBy class, but got stuck since the runtime type is specified when calling #callFunction and not when instantiating the class. It's really hard to explain why that was a problem, but I got around it by introducing an Aggregator class that could be instantiated in #callFunction where the type is known.