substring_*() functions can deal with any value as input
Admitted, this currently looks odd that some of the functions work on strings
only while others can accept any value. Whenever I could I relaxed the type
constraints on input in order for these functions to behave as much as their
XPath counterparts but maybe it does not so much value to worth it.

It is up for discussion.