Wrap arrays in List wrappers instead of creating copies in Adapter#toList
This is a neat performance improvement with no real cost. Instead of creating a new list and adding all elements to it, a new list wrapper is created that accesses the underlying array object directly.

I haven't figured out if it's possible to do the same for the object-to-list case.